fix(admin): invalid FILTER syntax broke the overview stats query
The single round-trip stats query (6f41196) wrote aggregates as
`count(*)::int FILTER (WHERE …)`, but Postgres requires FILTER to
attach directly to the aggregate — the cast in between is a syntax
error, so every /admin load failed with "Couldn't load this view"
once the refactor reached production. Move the cast outside:
`(count(*) FILTER (WHERE …))::int`. Verified against the production
database.

+++ src/app/admin/page.tsx
@@ -56,26 +56,27 @@ async function getStats() {
   const result = await db.execute(sql`
     SELECT
       (SELECT count(*)::int FROM employment_applications) AS apps_total,
-      (SELECT count(*)::int FILTER (WHERE status = 'new')
+      (SELECT (count(*) FILTER (WHERE status = 'new'))::int
          FROM employment_applications) AS apps_new,
       (SELECT count(*)::int FROM clinic_signups) AS clinic_leads_total,
-      (SELECT count(*)::int FILTER (WHERE status = 'new')
+      (SELECT (count(*) FILTER (WHERE status = 'new'))::int
          FROM clinic_signups) AS clinic_leads_new,
-      (SELECT count(*)::int FILTER (WHERE onboarding_completed)
+      (SELECT (count(*) FILTER (WHERE onboarding_completed))::int
          FROM clinics) AS accounts_total,
-      (SELECT count(*)::int FILTER (
-         WHERE onboarding_completed AND verification_status = 'pending')
+      (SELECT (count(*) FILTER (
+         WHERE onboarding_completed AND verification_status = 'pending'))::int
          FROM clinics) AS accounts_pending,
       (SELECT count(*)::int FROM email_signups) AS emails_total,
-      (SELECT count(*)::int FILTER (WHERE active) FROM promotions) AS promo_active,
-      (SELECT count(*)::int FILTER (WHERE active)
+      (SELECT (count(*) FILTER (WHERE active))::int
+         FROM promotions) AS promo_active,
+      (SELECT (count(*) FILTER (WHERE active))::int
          FROM featured_products) AS featured_active,
       (SELECT count(*)::int FROM pricing_quotes) AS quotes_total,
-      (SELECT count(*)::int FILTER (
+      (SELECT (count(*) FILTER (
          WHERE status = 'active'
-           AND (expires_at IS NULL OR expires_at > now()))
+           AND (expires_at IS NULL OR expires_at > now())))::int
          FROM pricing_quotes) AS quotes_active,
-      (SELECT count(*)::int FILTER (WHERE status = 'pending')
+      (SELECT (count(*) FILTER (WHERE status = 'pending'))::int
          FROM pricing_requests) AS pricing_requests_pending
   `);
